Fix light beam color on projected_spotlight.x3dv.
Previous way was a little hacky. It worked fine with per-vertex light, failed with per-pixel light.

- Previously, we had normal material (with white diffuse color), that was lit with spot light (color 1 0.75 0.75). The top of the light cone was light, the bottom was dark (as it had edges almost parallel to the light direction).

For per-vertex light, we could cheat: the light cone is actually quite tall, going significantly under the room floor. Since we interpolate lighting linearly, it's not visible that the bottom of light cone is dark. By stretching the cone, we stretch the gradient, and the top cone part (inside the room) seems uniformly light.

- For per-pixel light (used by shaders), this fails. Stretching the light cone doesn't change the gradient at the top, at the light cone is visibly darker.

The proper solution is just to use emissiveColor 1 0.75 0.75 for light cone. This is what we actually want.


git-svn-id: https://svn.code.sf.net/p/castle-engine/code/trunk/demo_models@8604 afe8dcc5-9094-451b-8773-5f495042a737