Make the smoothing algorithm (using creaseAngle) fair, independent of faces order. Fixes bug https://sourceforge.net/p/castle-engine/tickets/12/ . Adding testcases vrml_2/smoothing_independent_of_faces_order[-min].wrl .
Details: our previous algorithm was trying to be strict: for each vertex on each face, it was constructing a smoothing group containing only faces within the same creaseAngle. New faces were added to the smoothing group only if 1) they could be smoothed with *all* previous faces in the same group and 2) they were not yet part of any other smoothing group.

The implementation was done by FaceCanBeSmoothedWithFaces, and tracking was done by 1) SmoothFaces (to group strictly) and 2) HandledFaces (to make sure groups are disjoint) lists.

The advantage was that this was a careful (strict) approach: it made sure that all faces within a smoothing group satisfy the creaseAngle condition, that is: within a SmoothFaces group, every possible pair satisfies the creaseAngle condition. So every face in a group can be smoothed with every other face in a group.

The trouble with doing it this way is that order of faces determines the result. Once a face is added to some smoothing group, it limits what other faces can be added to the same smoothing group. The effect of this is visible in https://sourceforge.net/p/castle-engine/tickets/12/ testcase.

Both 1) and 2) conditions must be removed to make it really fair, common sense says this, and I also checked that merely removing 2) (removing HandledFaces variable) is not enough. Removing 2) helps, but it is not a 100% solution (it looks different, better, but still some shapes of be4oteo6.wrl are inconsistent). It's more-or-less obvious that removing 1) is not correct (consider a cone). So both 1) and 2) must be removed.

The new implementation: for each vertex, for each face: start by considering current face "initial". For all other faces: add them to a smoothing group if they can be smoothed with initial face. Don't care how the faces within a smoothing group relate to each other, the only important thing is how they relate to initial face. The advantage is that it's fair: only initial face determines what other faces are smoothed with it. The order of faces doesn't matter.

Also, smoothing group determined this way are not disjoint. In fact, we don't need to track smoothing groups at all, no need to track SmoothFaces or HandledFaces variables. So a nice side-effect is that normals calculation is now both simpler and faster.


git-svn-id: https://svn.code.sf.net/p/castle-engine/code/trunk/demo_models@12810 afe8dcc5-9094-451b-8773-5f495042a737